requirements.txt
deepdiff==6.3.0
prettytable==3.6.0
PyYAML==6.0
xmltodict==0.13.0
